move shouldRender to nextjs app
