+++ ws-nextjs-app/pages/[service]/articles/shouldRender/index.js
@@ -1,18 +1,10 @@
-import {
-  getPassportHome,
-  isValidPassportHome,
-  getCanonicalUrl,
-  matchesCanonicalUrl,
-} from '#lib/utilities/passport';
+import { getPassportHome, isValidPassportHome } from '#lib/utilities/passport';
 import { OK, NOT_FOUND } from '#lib/statusCodes.const';
-import { ARTICLE_PAGE } from '../../../../../routes/utils/pageTypes';
 
 // checks for pageData, 200 status and if home service from article data fits the service locale
 const shouldRender = (
   { pageData, status },
   service,
-  pathName,
-  pageType,
   passportHomesOverride = [],
 ) => {
   let statusCode = status;
@@ -24,18 +16,8 @@ const shouldRender = (
       service,
       passportHomesOverride,
     );
-    const isValidArticle = () => {
-      // Only check against Optimo Article pages
-      if (pageType !== ARTICLE_PAGE) return true;
-      if (service === 'sport') {
-        const canonicalUrl = getCanonicalUrl(pageData);
-        return matchesCanonicalUrl(canonicalUrl, pathName);
-      }
-      return true;
-    };
 
-    const isValidRequest = isValidService && isValidArticle();
-    statusCode = isValidRequest ? status : NOT_FOUND;
+    statusCode = isValidService ? status : NOT_FOUND;
   }
   const hasRequestSucceeded = hasDataAnd200Status && statusCode !== NOT_FOUND;
   return {
